Refactor file structure (for styled-components)
The file structure was designed for CSS Modules, which is not the type
of styling I'm using.
