+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.scss'
 
-import App from './components/App/App'
+import App from './components/App'
 import { HashRouter } from 'react-router-dom'
 
 const appJsx = (
